Make the title and logo templated
